refactor : spotify.service.js 파일명 변경 및 스웨거 에러코드 수정
